refactor: proxies folder.
