+++ src/common/index.ts
@@ -1,5 +1,4 @@
 export * from './modules';
-export * from './proxies';
 export * from './polkadot-provider';
 export * from './interceptors';
 export * from './health-indicators';
