+++ src/common/modules/substrate/substrate.service.ts
@@ -1,6 +1,6 @@
 import { ApiPromise, Keyring, WsProvider } from '@polkadot/api';
 import { Injectable, OnModuleInit, Logger } from '@nestjs/common';
-import { ProcessEnvProxy } from '../../proxies';
+import { ProcessEnvProxy } from '../..';
 
 @Injectable()
 export class SubstrateService implements OnModuleInit {
