+++ src/endpoints/verification/verification.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { DateTimeProxy } from '../../common/proxies/date-time';
+import { DateTimeProxy } from '../../common';
 import { RewardDto } from '../../common/modules/reward/dto/reward.dto';
 import { RewardService } from '../../common/modules/reward/reward.service';
 import {
