+++ test/unit/endpoints/verification/verification.service.spec.ts
@@ -1,11 +1,10 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { dateTimeProxyMockFactory, MockType } from '../../mock';
 import { when } from 'jest-when';
-import { DateTimeProxy } from '../../../../src/common/proxies/date-time';
 import { RewardService } from '../../../../src/common/modules/reward/reward.service';
 import { VerificationService } from '../../../../src/endpoints/verification/verification.service';
 import { RewardDto } from '../../../../src/common/modules/reward/dto/reward.dto';
-import { SubstrateService } from '../../../../src/common';
+import { DateTimeProxy, SubstrateService } from '../../../../src/common';
 import {
   convertToDbioUnitString,
   sendRewards,
